+++ adminforth/spa/src/afcl/Dropzone.vue
@@ -50,11 +50,13 @@
 import { humanifySize } from '@/utils';
 import { ref } from 'vue';
 import { IconFileSolid } from '@iconify-prerendered/vue-flowbite';
+import { watch } from 'vue';
 
 const props = defineProps<{
   extensions: string[],
   maxSizeBytes: number,
   multiple: boolean,
+  modelValue: FileList,
 }>();
 
 const emit = defineEmits(['update:modelValue']);
@@ -66,6 +68,14 @@ const selectedFiles: Ref<{
   size: number,
   mime: string,
 }[]> = ref([]);
+
+watch(() => props.modelValue, (files) => {
+  selectedFiles.value = Array.from(files).map(file => ({
+    name: file.name,
+    size: file.size,
+    mime: file.type,
+  }));
+});
 
 function doEmit(filesIn: FileList) {
 
@@ -81,8 +91,9 @@ function doEmit(filesIn: FileList) {
     mime: file.type,
   }));
 
-  emit('select', Array.from(files));
+  emit('update:modelValue', Array.from(files));
 }
 
 const dragging = ref(false);
+
 </script>