+++ adminforth/types/Back.ts
@@ -252,6 +252,7 @@ export interface IAdminForthDataSourceConnectorConstructor {
 
 export interface IAdminForthAuth {
   verify(jwt : string, mustHaveType: string, decodeUser?: boolean): Promise<any>;
+
   issueJWT(payload: Object, type: string, expiresIn?: string): string;
 
   removeCustomCookie({response, name}: {response: any, name: string}): void;
@@ -259,6 +260,8 @@ export interface IAdminForthAuth {
   setAuthCookie({expireInDays, response, username, pk,}: {expireInDays?: number, response: any, username: string, pk: string}): void;
   
   removeAuthCookie(response: any): void;
+
+  getClientIp(headers: any): string;
 }
 
 export interface IAdminForthRestAPI {
@@ -298,15 +301,15 @@ export interface IAdminForth {
   };
 
   createResourceRecord(
-    params: { resource: AdminForthResource, record: any, adminUser: AdminUser }
+    params: { resource: AdminForthResource, record: any, adminUser: AdminUser, extra?: HttpExtra }
   ): Promise<{ error?: string, createdRecord?: any }>;
 
   updateResourceRecord(
-    params: { resource: AdminForthResource, recordId: any, record: any, oldRecord: any, adminUser: AdminUser }
+    params: { resource: AdminForthResource, recordId: any, record: any, oldRecord: any, adminUser: AdminUser, extra?: HttpExtra }
   ): Promise<{ error?: string }>;
 
   deleteResourceRecord(
-    params: { resource: AdminForthResource, recordId: string, adminUser: AdminUser, record: any }
+    params: { resource: AdminForthResource, recordId: string, adminUser: AdminUser, record: any, extra?: HttpExtra }
   ): Promise<{ error?: string }>;
 
   auth: IAdminForthAuth;
@@ -434,6 +437,12 @@ export type AfterDataSourceResponseFunction = (params: {
   adminforth: IAdminForth,
 }) => Promise<{ok: boolean, error?: string}>;
 
+export interface HttpExtra {
+  body: any,
+  query: Record<string, string>,
+  headers: Record<string, string>,
+  cookies: Record<string, string>,
+}
 /**
  * Modify record to change how data is saved to database.
  * Return ok: false and error: string to stop execution and show error message to user. Return ok: true to continue execution.
@@ -445,6 +454,7 @@ export type BeforeSaveFunction = (params: {
   record: any, 
   oldRecord?: any,
   adminforth: IAdminForth,
+  extra?: HttpExtra,
 }) => Promise<{ok: boolean, error?: string}>;
 
 /**
@@ -458,6 +468,7 @@ export type AfterSaveFunction = (params: {
   record: any, 
   oldRecord?: any,
   adminforth: IAdminForth,
+  extra?: HttpExtra,
 }) => Promise<{ok: boolean, error?: string}>;
 
 /**
@@ -467,6 +478,7 @@ export type BeforeLoginConfirmationFunction = (params?: {
     adminUser: AdminUser,
     response: IAdminForthHttpResponse,
     adminforth: IAdminForth,
+    extra?: HttpExtra,
 }) => Promise<{
   error?: string, 
   body: {
@@ -840,6 +852,13 @@ export interface AdminForthInputConfig {
        * @returns 
        */
       websocketSubscribed?: (topic: string, user: AdminUser) => void,
+
+      /**
+       * Client IP header name. If set, AdminForth will use this header to get client IP address.
+       * Otherwise it will use first IP address from X-Forwarded-For header.
+       * If you are using Cloudflare, set this to 'CF-Connecting-IP'. Case-insensitive.
+       */
+      clientIpHeader?: string,
     },
 
      /**